+++ src/server/src/models/services/folsService.py
@@ -233,11 +233,3 @@ def registerAccess(folTitle, position, userId):
         folAccessAttempt["userName"] = user["Username"]
 
     conn.insert("folconn", "folAccessAttempts", folAccessAttempt)
-
-def getFolAcessesByUserName(user):
-    conn = PyMongoConnection()
-    documents = jsonify(list(conn.getDocuments("folconn", "folAccessAttempts", {"userName":user}, {"_id": 0, "id": 0})))
-
-    response = make_response(documents)
-
-    return response
